requirements.txt
absl-py==1.4.0
aniso8601==9.0.1
astor==0.8.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cycler==0.11.0
Flask==1.0.2
Flask-RESTful==0.3.7
fonttools==4.38.0
gast==0.5.4
gevent==1.4.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
gunicorn==20.0.4
h5py==3.8.0
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.2.0
networkx==2.6.3
numpy==1.16.4
onnxruntime-gpu==1.1.0
opencv-python==4.2.0.32
packaging==23.1
Pillow==9.5.0
pip==22.3.1
protobuf==3.19.0
py-eureka-client==0.5.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.0.0
requests==2.30.0
scikit-image==0.16.2
scipy==1.5.0
setuptools==65.6.3
six==1.16.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==2.3.0
tifffile==2021.11.2
torch==1.11.0
torchvision==0.12.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.2.3
wheel==0.38.4
wrapt==1.15.0
zipp==3.15.0
